requirements.txt
babel==2.17.0
defusedxml==0.7.1
et_xmlfile==2.0.0
fonttools==4.61.1
fpdf2==2.8.5
numpy==2.4.1
openpyxl==3.1.5
pandas==2.3.3
pillow==12.1.0
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
tkcalendar==1.6.1
tzdata==2025.3
